+++ src/main/java/com/d/utegilishev/TelegramBotwithChatGPT/TelegramBot.java
@@ -5,12 +5,16 @@ import org.telegram.telegrambots.bots.DefaultBotOptions;
 import org.telegram.telegrambots.bots.TelegramLongPollingBot;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Update;
+import com.d.utegilishev.TelegramBotwithChatGPT.openai.aiClient;
 
 public class TelegramBot extends TelegramLongPollingBot {
 
+    private final aiClient openAIClient;
 
-    public TelegramBot(DefaultBotOptions options, String botToken) {
+    public TelegramBot(DefaultBotOptions options, String botToken, aiClient openAIClient ) {
         super(options, botToken);
+        this.openAIClient = openAIClient;
+
     }
 
 
@@ -20,7 +24,10 @@ public class TelegramBot extends TelegramLongPollingBot {
         if (update.hasMessage() && update.getMessage().hasText()) {
             var text = update.getMessage().getText();
             var chatId = update.getMessage().getChatId();
-            SendMessage sendMessage = new SendMessage(chatId.toString(), text);
+            //implement response from gpt
+            var chatComplitionResponse = openAIClient.createChatComplition(text);
+            var textResponse = chatComplitionResponse.getChoices().get(0).getMessage().getContent();
+            SendMessage sendMessage = new SendMessage(chatId.toString(), textResponse);
             sendApiMethod(sendMessage);
         }
     }