+++ src/main/java/com/d/utegilishev/TelegramBotwithChatGPT/openai/Choice.java
@@ -1,0 +1,11 @@
+package com.d.utegilishev.TelegramBotwithChatGPT.openai;
+
+
+
+import lombok.Data;
+
+@Data
+public class Choice {
+
+    private final Message message;
+}