+++ src/main/java/com/d/utegilishev/TelegramBotwithChatGPT/openai/Message.java
@@ -1,0 +1,12 @@
+package com.d.utegilishev.TelegramBotwithChatGPT.openai;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+public class Message {
+
+    private final String role;
+    private final String content;
+
+}