+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ### Version 0.0.1
 - start project, add dependency and creating bot itself
-- 
+- reading the documentation of openAI
+- Main data got in the json filed choices
+### version 0.0.2
+- create DTO for reading json response from gpt
+- add restTemplate builder for building http request 
+- add openAi client for creating request and response
+- need to test dto etc